Upgrade QGIS core to 3.40.6

+++ src/3d/qgs3daxissettings.cpp
@@ -32,7 +32,7 @@ bool Qgs3DAxisSettings::operator==( Qgs3DAxisSettings const &rhs ) const
 
 bool Qgs3DAxisSettings::operator!=( Qgs3DAxisSettings const &rhs ) const
 {
-  return ! this->operator==( rhs );
+  return !this->operator==( rhs );
 }
 
 void Qgs3DAxisSettings::readXml( const QDomElement &element, const QgsReadWriteContext & )
@@ -43,11 +43,11 @@ void Qgs3DAxisSettings::readXml( const QDomElement &element, const QgsReadWriteC
 
   sizeStr = element.attribute( QStringLiteral( "minViewportRatio" ) );
   if ( !sizeStr.isEmpty() )
-    mMinViewportRatio = sizeStr.toInt();
+    mMinViewportRatio = sizeStr.toDouble();
 
   sizeStr = element.attribute( QStringLiteral( "maxViewportRatio" ) );
   if ( !sizeStr.isEmpty() )
-    mMaxViewportRatio = sizeStr.toInt();
+    mMaxViewportRatio = sizeStr.toDouble();
 
   const QString modeStr = element.attribute( QStringLiteral( "mode" ) );
   if ( modeStr == QLatin1String( "Off" ) )
@@ -132,7 +132,6 @@ void Qgs3DAxisSettings::writeXml( QDomElement &element, const QgsReadWriteContex
       break;
   }
   element.setAttribute( QStringLiteral( "vertical" ), str );
-
 }
 
 void Qgs3DAxisSettings::setMinViewportRatio( double ratio )
